:wrench: config: Update color palette (#7)
Currently the contrast between the colors is not good enough and they
are not going well together. So, I updated the color's here to improve
the User Interface. Also, this changes are from the theme repo, which
makes it easier to maintain this repo in future.

:package: theme: Update git submodule to latest upstream changes

Adds needed changes that make the color palette change noticeable.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>