:package: :arrow_up: theme: Upstream layout changes
This commit updates the theme submodule with new commits added upstream
by @Neha9849. The most notable change is to the sidebar navigation and
the breadcrumb bar. The design is now more responsive and works better
on mobile devices. Additionally, the Team Profiles / Cohorts feature now
uses the same code as other parts of the site, in order to reduce the
maintenance burden.

Excerpt of commits included in this update:

* :recycle: layouts: Unify cohorts partials with theme layout
* :lipstick: layouts: Merge breadcrumb bar to top, improve sidebar navigation
* :recycle: breadcrumbs: Use Hugo-native script instead of JavaScript
* :lipstick: assets: Fix footer icons and URLs

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>